fix: subarray offset bug
