+++ qoi/deno.json
@@ -1,6 +1,6 @@
 {
   "name": "@img/qoi",
-  "version": "0.1.3",
+  "version": "0.1.4",
   "exports": {
     ".": "./mod.ts",
     "./decode": "./decode.ts",
