+++ qoi/encoder_stream.ts
@@ -1,4 +1,4 @@
-import { createEncoder } from "./_common.ts";
+import { createEncoder } from "./_encoder.ts";
 import type { QOIOptions } from "./types.ts";
 
 /**
@@ -60,6 +60,11 @@ export class QOIEncoderStream extends TransformStream<Uint8Array, Uint8Array> {
       transform(chunk, controller): void {
         const originalSize = chunk.length;
         const maxSize = Math.ceil(offset + originalSize / 4) * (isRGB ? 4 : 5);
+        if (chunk.byteOffset) {
+          const buffer = new Uint8Array(chunk.buffer);
+          buffer.set(chunk);
+          chunk = buffer.subarray(0, chunk.length);
+        }
         // deno-lint-ignore no-explicit-any
         chunk = new Uint8Array((chunk.buffer as any).transfer(maxSize));
         chunk.set(chunk.subarray(0, originalSize), maxSize - originalSize);
